% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{aao_dprior_tau}
\alias{aao_dprior_tau}
\title{(Log) density of the prior distribution of the variance scaling constant}
\usage{
aao_dprior_tau(tau, v)
}
\arguments{
\item{tau}{The covariance scalar}

\item{v}{The variance hyperparameter}
}
\value{
The prior log density for tau
}
\description{
The covariance scalar needs to be estimated as well. It can be interpreted as a
penalty term. The currently implemented prior is a half-normal distribution
with variance hyperparameter v.
}
\examples{
aao_dprior_tau(1, 0.1)
aao_dprior_tau(0, 0.1) # returns -Inf
}
