% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{aao_rq_beta}
\alias{aao_rq_beta}
\alias{aao_dq_beta}
\title{The proposal distribution for the spline coefficients}
\usage{
aao_rq_beta(center, tau)

aao_dq_beta(beta, center, tau)
}
\arguments{
\item{center}{The mean of the proposal distribution}

\item{tau}{The variance hyperparameter}

\item{beta}{The spline coefficients}
}
\value{
For \code{aao_rq_beta}: A new set of proposed spline coefficients

For \code{aao_dq_beta}: The log density of beta, with respect to a
  distribution with mean \code{center}
}
\description{
Currently this is implemented as a \code{k}-dimensional multivariate normal
distribution. The covariance matrix is that of a
penalty matrix of second-order differences, as described in Lang & Brezger
(2001).
}
\examples{
# proposal function
beta <- c(-6.5188650, 1.39889900,-0.2601626,
          -1.2675330, 0.04916187, 1.9759470,
           1.3406900,-2.23430300,-2.1876150,
           1.1648310, 2.01463700, 0.9876818,
          -0.4144843,-1.35885900,-4.1775160)
tau <- 0.1
aao_rq_beta(beta, tau)


# density function
beta <- c(-6.5188650, 1.39889900,-0.2601626,
          -1.2675330, 0.04916187, 1.9759470,
           1.3406900,-2.23430300,-2.1876150,
           1.1648310, 2.01463700, 0.9876818,
          -0.4144843,-1.35885900,-4.1775160)
tau <- 0.1
beta2 <- aao_rq_beta(beta, tau)
aao_dq_beta(beta, beta2, tau)
}
