% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{aao_rq_tau}
\alias{aao_rq_tau}
\alias{aao_dq_tau}
\title{The proposal distribution for the covariance scalar}
\usage{
aao_rq_tau(center, v)

aao_dq_tau(tau, center, v)
}
\arguments{
\item{center}{The mean of the proposal distribution}

\item{v}{The variance hyperparameter}

\item{tau}{The covariance scalar}
}
\value{
For \code{aao_rq_tau}: A new proposed covariance scalar

For \code{aao_dq_tau}: The log density of tau, with respect to a
  distribution with mean \code{center}
}
\description{
The proposal distribution for the covariance scalar
}
\examples{
# proposal function
aao_rq_tau(0.5, 0.1)


# density function
aao_dq_tau(0.5, 0.4, 0.1)
}
