% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{buildmat}
\alias{buildmat}
\title{Prior covariance matrix for the spline coefficients}
\usage{
buildmat(n)
}
\arguments{
\item{n}{The number of spline coefficients}
}
\value{
An n x n penalty matrix
}
\description{
This is a penalty matrix of second-order differences, as described in Lang &
Brezger (2001).
}
