% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{covar_plot}
\alias{covar_plot}
\title{Plot the covariance function after estimation}
\usage{
covar_plot(h, beta, knots, gpModel = NULL, beta_true = NULL,
  beta_init = NULL, err_bounds = NULL)
}
\arguments{
\item{h}{The lag values to plot the covariance function over}

\item{beta}{A vector of spline coefficients (typically the final result from
\code{\link{estbeta}})}

\item{knots}{The knot locations}

\item{gpModel}{A \code{GPmodel} object}

\item{beta_true}{The "true" spline coefficients}

\item{beta_init}{The initial spline coefficient values}

\item{err_bounds}{A 2-column matrix containing lower and upper error bounds}
}
\value{
The \code{GPmodel} object, invisibly. Creates the plot as a side
  effect.
}
\description{
Plot the covariance function after estimation
}
\details{
All of \code{gpModel}, \code{beta_true}, \code{beta_init}, and
  \code{err_bounds} are \code{NULL} by default. If any are not known or not
  desired on the final plot, leave them as \code{NULL}. Otherwise, explicitly
  provide them.
}
