% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{create_locations}
\alias{create_locations}
\title{Create observation and prediction locations for simulating a Gaussian process}
\usage{
create_locations(M, ngrid, mindist = 0.005)
}
\arguments{
\item{M}{The number of observation locations}

\item{ngrid}{The prediction grid will be \code{ngrid} x \code{ngrid} over the
unit square.}

\item{mindist}{The minimum possible distance between any two observations. If
two observations are too close together, it may cause numerical
instability.}
}
\value{
A \code{GPlocations} object containing the following:
  \itemize{
    \item \code{locs}: a data frame with three columns - \code{x} and \code{y}
    are the coordinates of the locations, and \code{type} is a factor equal to
    \code{obs} for an observation location and \code{pred} for a prediction
    location.
    \item \code{dist_obs}: an \code{M} x \code{M} matrix consisting of the
    distances between the observation locations
    \item \code{dist_pred}: an \code{ngrid} x \code{ngrid} matrix consisting
    of the distances between the prediction locations
    \item \code{mindist}: The \code{mindist} parameter for reference
  }
}
\description{
Generates a \code{GPlocations} object, containing locations for both
(non-gridded) observations and (gridded) predictions when the Gaussian
process is simulated
}
\examples{
locations <- create_locations(50, 20, mindist = 0.002)
}
