% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{create_locations2}
\alias{create_locations2}
\title{Create observation and prediction locations for simulating a Gaussian process}
\usage{
create_locations2(M, grid_pred, mindist = 0.005)
}
\arguments{
\item{M}{The number of observation locations}

\item{grid_pred}{A data frame with two columns, \code{x} and \code{y}, that
contains the prediction points}

\item{mindist}{The minimum possible distance between any two observations. If
two observations are too close together, it may cause numerical
instability.}
}
\value{
A \code{GPlocations} object containing the following: \itemize{ \item
  \code{locs}: a data frame with three columns - \code{x} and \code{y} are
  the coordinates of the locations, and \code{type} is a factor equal to
  \code{obs} for an observation location and \code{pred} for a prediction
  location. \item \code{dist_obs}: an \code{M} x \code{M} matrix consisting
  of the distances between the observation locations \item \code{dist_pred}:
  an \code{nrow(grid_pred)} x \code{nrow(grid_pred)} matrix consisting of the
  distances between the prediction locations \item \code{mindist}: The
  \code{mindist} parameter for reference }
}
\description{
Generates a \code{GPlocations} object. This is an altered version of
\code{\link{create_locations}}, in which the observations are restricted to
fall outside of (1/3, 2/3) x (1/3, 2/3) and the prediction points are
specified in an argument. The prediction points should be limited to the
square that the observation points have been withheld from.
}
\examples{
locations <- create_locations2(50, data.frame(x = 0.5, y = 0.5), mindist = 0.002)
}
