% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{d_prior_beta}
\alias{d_prior_beta}
\title{(Log) density of the joint prior distribution on the spline coefficients}
\usage{
d_prior_beta(beta, tau)
}
\arguments{
\item{beta}{The spline coefficients}

\item{tau}{A constant that scales the covariance matrix (also will be
estimated by the MCMC)}
}
\value{
The prior log density for beta
}
\description{
This is a \code{k}-dimensional multivariate normal distribution, where
\code{k} is the dimension of \code{beta}. The covariance is currently set to
be the same as the proposal covariance (see \code{\link{d_q_beta}}).
}
\examples{
beta <- c(-6.5188650, 1.39889900,-0.2601626,
          -1.2675330, 0.04916187, 1.9759470,
           1.3406900,-2.23430300,-2.1876150,
           1.1648310, 2.01463700, 0.9876818,
          -0.4144843,-1.35885900,-4.1775160)
tau <- 0.01
d_prior_beta(beta, tau)
}
