% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dloglogspline}
\alias{dloglogspline}
\title{Evaluate the log of the log spectral density}
\usage{
dloglogspline(x, coefs, knots)
}
\arguments{
\item{x}{The values at which to evaluate the log of the log density.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of log log density values.
}
\description{
Evaluate the log of the log spectral density
}
