% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlogspline}
\alias{dlogspline}
\title{Evaluate the log spectral density}
\usage{
dlogspline(x, coefs, knots)
}
\arguments{
\item{x}{The values at which to evaluate the log density.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of log density values.
}
\description{
Evaluate the log spectral density
}
