% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlogspline_unnorm}
\alias{dlogspline_unnorm}
\title{Evaluate the un-normalized log spectral density}
\usage{
dlogspline_unnorm(basis, coefs, knots)
}
\arguments{
\item{basis}{An \code{n} by \code{k} spline basis matrix, where \code{n}
is the number of points the curve will be evaluated at and \code{k} is the
number of knots. Should be the output of \code{\link{nsbasis}}.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of length \code{n} containing the log density.
}
\description{
Evaluate the un-normalized log spectral density
}
\seealso{
\code{\link{predict_natspl}} for the log of this function.
}
