% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_batch.R
\name{estbeta_initialize}
\alias{estbeta_initialize}
\title{Estimate spline coefficients (initializer)}
\usage{
estbeta_initialize(filename, errname, nbatch, H, Y, B, init_beta, init_tau,
  knots, v_beta, v_tau, t_v, r.opt = 0.23, r = 5, eps = 0.001,
  nugget = 0.001)
}
\arguments{
\item{filename}{Name of the file to store the betas in}

\item{errname}{Name of the file to store the covariance functions in}

\item{nbatch}{Number of iterations to run this function for. See "Details"}

\item{H}{The distance matrix between the observations. \code{M} x \code{M}
matrix, where \code{M} is the number of observations.}

\item{Y}{The vector of observations. Length \code{M}}

\item{B}{The number of random draws to take from the spline-approximated
spectral density}

\item{init_beta}{Initial values for the spline coefficients}

\item{init_tau}{Initial value for the variance scalar}

\item{knots}{The vector of knot locations}

\item{v_beta}{The \code{beta} proposal variance hyperparameter}

\item{v_tau}{The \code{tau} proposal variance hyperparameter}

\item{t_v}{The prior variance hyperparameter (not being adapted)}

\item{r.opt}{The optimal acceptance rate (used for adaptive tuning)}

\item{r}{The number of previous iterations to use for adaptive tuning}

\item{eps}{Not used I think}

\item{nugget}{Nugget effect}
}
\value{
Returns a list of information to pass to \code{\link{estbeta}}. Also writes output to some csv files.
}
\description{
This and \code{\link{estbeta}} probably should not be two separate functions
but oh well. This function contains the main MCMC loop.
}
\details{
If \code{nbatch} is too big, the program will crash eventually
  because R is dumb and won't release the memory after each expensive
  likelihood calculation. A value of around 100 is recommended.
}
