% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{find_max_ratio}
\alias{find_max_ratio}
\title{Find max ratio}
\usage{
find_max_ratio(nu, alpha, sigma, df)
}
\arguments{
\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{df}{degrees of freedom of the t distribution}
}
\value{
the maximum value of f(x), where f is the ratio of the Matern to the t
}
\description{
Find the maximum of \code{rat_matern_t}, the ratio of the Matern density to the t density.
}
\seealso{
\code{\link{rat_matern_t}}
}
