% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{fit_matern}
\alias{fit_matern}
\title{Fit a Matern model to a GPsimulated object}
\usage{
fit_matern(gpEst, n_samps, starting = list(sigma.sq = 1, tau.sq = 0.1, phi =
  0.5, nu = 1), tuning = list(sigma.sq = 0.01, tau.sq = 0.01, phi = 0.005, nu
  = 0.005), priors = list(sigma.sq.ig = c(2, 1), tau.sq.ig = c(2, 1), phi.unif
  = c(1e-05, 20), nu.unif = c(1e-05, 5)))
}
\arguments{
\item{gpEst}{A \code{GPsimulated} object}

\item{n_samps}{Number of MCMC samples to take}

\item{starting}{Starting parameter values}

\item{tuning}{Tuning parameter values}

\item{priors}{Priors and hyperparameters}
}
\value{
A list containing the best-fitting Matern model, the result of \code{\link[spBayes]{spLM}}, and the parameters (transformed back into our parameterization)
}
\description{
Uses \code{\link[spBayes]{spLM}} to fit a Matern model to the data.
}
\details{
See \code{\link[spBayes]{spLM}} for details on the arguments. Some sensible defaults are provided but may need to be tweaked.
}
