% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{fitspline}
\alias{fitspline}
\title{Bayesian spline fitting}
\usage{
fitspline(B, y, spl, burnin = floor(B/5), init = list(beta = rep(1,
  ncol(spl)), sigma = 1, tau = 1), tune_init = list(beta = rep(1, ncol(spl)),
  sigma = 1, tau = 1), hyperparams = list(sigma = c(1, 0.005), tau = c(1,
  0.005)), c0 = 10, c1 = 0.8, k = 3, r = 5, opt.acc.rate = 0.41,
  progress = T)
}
\arguments{
\item{B}{the number of MCMC iterations}

\item{y}{a vector of observations}

\item{spl}{a natural cubic spline basis matrix, such as one generated by
\code{\link{nsbasis}}}

\item{burnin}{the number of iterations to discard as burn-in}

\item{init}{a named list of initial guesses for \code{beta} (the basis
coefficients), \code{sigma} (the model error estimate), and \code{tau} (the
smoothness penalty).}

\item{tune_init}{a named list of initial guesses for tuning parameters}

\item{hyperparams}{a named list of hyperparameters for the priors on
\code{sigma} and \code{tau}. Currently these priors are both hard coded as
inverse gamma.}

\item{c0}{adaptive tuning parameter}

\item{c1}{adaptive tuning parameter}

\item{k}{adaptive tuning parameter}

\item{r}{lag for adaptive tuning}

\item{opt.acc.rate}{optimal accept rate for adaptive tuning}

\item{progress}{not currently used}
}
\value{
A list of length 3. The first element is a matrix where each row is a
  random \code{beta} draw. The second and third are vectors where each
  element is a random draw of \code{sigma} and \code{tau} respectively.
}
\description{
Bayesian spline fitting
}
