% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitsplinecpp}
\alias{fitsplinecpp}
\title{Bayesian spline fitting, in C++}
\usage{
fitsplinecpp(B, y, spl, burnin, initBeta, initSigma, initTau, initBetaTune,
  initSigmaTune, initTauTune, s_a, s_b, t_a, t_b, c0, c1, k, r, oar, progress)
}
\arguments{
\item{B}{B}

\item{y}{y}

\item{spl}{spl}

\item{burnin}{burning}

\item{initBeta}{initBeta}

\item{initSigma}{initSigma}

\item{initTau}{initTau}

\item{initBetaTune}{initBetaTune}

\item{initSigmaTune}{initSigmaTune}

\item{initTauTune}{initTauTune}

\item{s_a}{s_a}

\item{s_b}{s_b}

\item{t_a}{t_a}

\item{t_b}{t_b}

\item{c0}{c0}

\item{c1}{c1}

\item{k}{k}

\item{r}{r}

\item{oar}{oar}

\item{progress}{progress}
}
\value{
A list of length 3. The first element is a matrix where each row is a
  random \code{beta} draw. The second and third are vectors where each
  element is a random draw of \code{sigma} and \code{tau} respectively.
}
\description{
Don't use this directly. Use the wrapper function \code{\link{fitspline}}.
}
