% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{generateGP}
\alias{generateGP}
\title{Generate a Gaussian process}
\usage{
generateGP(n, nu, alpha, sigma, tau, mindist = 0.005)
}
\arguments{
\item{n}{the number of observations to generate}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{tau}{nugget effect}

\item{mindist}{the minimum distance between two observation locations}
}
\value{
a list with two elements. The first is a data frame with the x and y
  coordinates of the observation locations and z, which is the value of the
  observation. The second is a matrix of distances between the observations.
  We return this so that it doesn't need to be calculated separately later
  on.
}
\description{
Generates a Gaussian random field on the unit square. Observation locations
are guaranteed to be a certain minimum distance apart.
}
