% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_slope_pts}
\alias{get_slope_pts}
\title{Evaluate points beyond the knot boundaries}
\usage{
get_slope_pts(coefs, knots)
}
\arguments{
\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of length 4 containing the following values:
\enumerate{
  \item \eqn{f(kmin - 1)}
  \item \eqn{f(kmin)}
  \item \eqn{f(kmax)}
  \item \eqn{f(kmax + 1)}
}
where \code{kmin} and \code{kmax} are the minimum and maximum knots and
\code{f} is the log log density. Then the left tail slope is
\eqn{f(kmin) - f(kmin - 1)} and the right tail slope is
\eqn{f(kmax+ 1) - f(kmax)}.
}
\description{
Since we are using a natural spline basis, the fitted curve beyond the
boundary knots is guaranteed to be linear. To quickly find the slopes
of these linear tails, this function evaluates the fitted curve at four
points: the boundary knots, as well as at a point slightly
beyond each boundary knot.
}
