% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_slopes}
\alias{get_slopes}
\title{Get slopes beyond the knot boundaries}
\usage{
get_slopes(coefs, knots)
}
\arguments{
\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of length 2 containing the following values:
\enumerate{
  \item \eqn{f(kmin) - f(kmin - 1)}
  \item \eqn{f(kmax + 1) - f(kmax)}
}
where \code{kmin} and \code{kmax} are the minimum and maximum knots and
\code{f} is the unnormalized log log density. The first value is the left
tail slope and the second is the right tail slope.
}
\description{
Get slopes beyond the knot boundaries
}
