% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{gp_mse_optimal}
\alias{gp_mse_optimal}
\title{Optimally predict points from a GPsimulated object and find the mean squared
error}
\usage{
gp_mse_optimal(gpEst)
}
\arguments{
\item{gpEst}{A \code{GPsimulated} object}
}
\value{
The mean squared error
}
\description{
If the predictions have already been found via \code{\link{optimal_predict}},
use \code{\link{mse}} instead.
}
