% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{gp_mse_othermodel}
\alias{gp_mse_othermodel}
\title{Optimally predict points using an arbitrary covariance model and find the
mean squared error}
\usage{
gp_mse_othermodel(gpEst, model)
}
\arguments{
\item{gpEst}{A \code{GPsimulated} object}

\item{model}{A covariance model made using the \code{RandomFields} package}
}
\value{
The mean squared error
}
\description{
To predict points from another model, such as the best-fitting Matern model,
create it using \code{RandomFields} and specify it here.
}
\details{
If the predictions have already been found via
\code{\link{predict.GPsimulated}}, use \code{\link{mse}} instead.
}
