% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{likelihood}
\alias{likelihood}
\title{Calculate the approximate likelihood from the spline coefficients}
\usage{
likelihood(beta, knots, Y, H, B, nugget, debug)
}
\arguments{
\item{beta}{The spline coefficients. Vector of length \code{k}}

\item{knots}{The knot locations. Vector of length \code{k}}

\item{Y}{The observations. Vector of length \code{M}}

\item{H}{The distance matrix between the observations. \code{M} x \code{M}
matrix}

\item{B}{The number of random draws to take from the spline-approximated
spectral density}

\item{nugget}{The nugget effect}

\item{debug}{\code{TRUE} to print out debugging information. (This is
untested inside the \code{gpcovr} package. Best to leave this
\code{FALSE}.)}
}
\value{
The approximate log likelihood
}
\description{
Calculate the approximate likelihood from the spline coefficients
}
