% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{loglogdensity_plot}
\alias{loglogdensity_plot}
\title{Plot the log-log spectral density after estimation}
\usage{
loglogdensity_plot(beta, gpModel, knots, show_knots = TRUE,
  beta_true = NULL, beta_init = NULL, err_bounds = NULL)
}
\arguments{
\item{beta}{A vector of spline coefficients (typically the final result from
\code{\link{estbeta}})}

\item{gpModel}{A \code{GPmodel} object}

\item{knots}{The knot locations}

\item{show_knots}{To hide the vertical dashed lines indicating where the
knots are located, set this to FALSE. Defaults to TRUE}

\item{beta_true}{The "true" spline coefficients}

\item{beta_init}{The initial spline coefficient values}

\item{err_bounds}{A 2-column matrix containing lower and upper error bounds}
}
\value{
The \code{GPmodel} object, invisibly. Creates the plot as a side effect.
}
\description{
Plot the log-log spectral density after estimation
}
\details{
All of \code{beta_true}, \code{beta_init}, and \code{err_bounds} are \code{NULL} by default. If any are not known or not desired on the final plot, leave them as \code{NULL}. Otherwise, explicitly provide them.
}
