% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{matern_cor}
\alias{matern_cor}
\title{Matern covariance function}
\usage{
matern_cor(h, nu, alpha, sigma)
}
\arguments{
\item{h}{distance value/vector/matrix}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}
}
\value{
The covariance associated with the distance h
}
\description{
The true Matern covariance function.
}
