% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(pred, actual)
}
\arguments{
\item{pred}{Predicted values}

\item{actual}{Actual values}
}
\value{
The mean squared error: \code{mean((pred - actual)^2)}
}
\description{
Mean squared error
}
