% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{normal_ll}
\alias{normal_ll}
\title{Approximated normal log likelihood, Matern covariance}
\usage{
normal_ll(x, h, nu, alpha, sigma, tau, num_samps, mc_samps = NULL)
}
\arguments{
\item{x}{vector of observations}

\item{h}{distance matrix between the observations}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{tau}{nugget effect}

\item{num_samps}{the number of MC samples to use. Ignored if \code{mc} is not null.}

\item{mc_samps}{vector of MC samples. If \code{NULL}, generate them with calls to \code{\link{rmatern}} and \code{cudatest::mc}.}
}
\value{
the log likelihood.
}
\description{
Approximated normal log likelihood, Matern covariance
}
