% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{normal_ll_exact}
\alias{normal_ll_exact}
\title{Exact normal log likelihood}
\usage{
normal_ll_exact(x, h, nu, alpha, sigma, tau)
}
\arguments{
\item{x}{vector of observations}

\item{h}{distance matrix between the observations}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{tau}{nugget effect}
}
\value{
the log likelihood.
}
\description{
Exact normal log likelihood
}
