% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_dlogspline}
\alias{normalize_dlogspline}
\title{Calculate the normalizing constant for the log spectral density}
\usage{
normalize_dlogspline(basis, coefs, knots)
}
\arguments{
\item{basis}{An \code{n} by \code{k} spline basis matrix, where \code{n}
is the number of points the curve will be evaluated at and \code{k} is the
number of knots. Should be the output of \code{\link{nsbasis}}.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
the normalizing constant C
}
\description{
The log density given by \code{\link{dlogspline_unnorm}} is not a valid
density because it does not generally integrate to 1. This function finds
the constant C such that 1/C times \code{dlogspine_unnorm} will integrate
to 1.
}
\seealso{
\code{\link{dlogspline_unnorm}}, \code{\link{dlogspline}}
}
