% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nsbasis}
\alias{nsbasis}
\title{Create a natural cubic spline basis}
\usage{
nsbasis(X, knots)
}
\arguments{
\item{X}{the locations at which the spline functions will be evaluated. The
matrix will have \code{length(X)} rows.}

\item{knots}{the knot locations. The first and last knots are the boundary
knots, and the rest are internal knots. The matrix will have
\code{length(knots)} knots.}
}
\value{
the basis matrix
}
\description{
Create a natural cubic spline basis
}
\details{
The basis used by this function is described in \emph{The Elements
of Statistical Learning} by Hastie and Tibshirani, in the appendix of
Chapter 5 (2nd edition).
}
