% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_ooat.R
\name{ooat_dprior_beta}
\alias{ooat_dprior_beta}
\title{(Log) density of the prior distribution on each spline coefficient}
\usage{
ooat_dprior_beta(beta, tau, i)
}
\arguments{
\item{beta}{The spline coefficients}

\item{tau}{The prior variance parameter}

\item{i}{The position of the coefficient within \code{beta}}
}
\value{
The prior log density for \code{beta[i]}
}
\description{
This is a normal prior on each of the coefficients. The means are second-order differences, see Lang & Brezger (2001).
}
\examples{
beta <- c(-6.5188650, 1.39889900,-0.2601626,
          -1.2675330, 0.04916187, 1.9759470,
           1.3406900,-2.23430300,-2.1876150,
           1.1648310, 2.01463700, 0.9876818,
          -0.4144843,-1.35885900,-4.1775160)
tau <- 0.01
ooat_dprior_beta(beta, tau, 4)
}
