% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_ooat.R
\name{ooat_rq_beta}
\alias{ooat_rq_beta}
\alias{ooat_dq_beta}
\title{The proposal distribution for each of the spline coefficients}
\usage{
ooat_rq_beta(center, v_beta)

ooat_dq_beta(beta, center, v_beta)
}
\arguments{
\item{center}{Mean of the proposal distribution}

\item{v_beta}{Variance of the proposal distribution}

\item{beta}{The current spline coefficient}
}
\value{
For \code{ooat_rq_beta}: A new proposed spline coefficient

For \code{ooat_dq_beta}: The log density of beta, with respect to a
  distribution with mean \code{center}.
}
\description{
This is just a normal distribution with mean \code{center} and variance
\code{v_beta}.
}
\examples{
# proposal function
ooat_rq_beta(2.2, 0.1)


# density function
ooat_dq_beta(2, 2.2, 0.1)
}
