% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_ooat.R
\name{ooat_rq_tau}
\alias{ooat_rq_tau}
\alias{ooat_dq_tau}
\title{The proposal distribution for the variance scalar}
\usage{
ooat_rq_tau(center, v)

ooat_dq_tau(tau, center, v)
}
\arguments{
\item{center}{The mean of the proposal distribution}

\item{v}{The variance of the proposal distribution}

\item{tau}{The current variance scalar}
}
\value{
For \code{ooat_rq_tau}: A new proposed variance scalar

For \code{ooat_dq_tau}: The log density of tau, with respect to a
  distribution with mean \code{center}
}
\description{
This is just a normal distribution with mean \code{center} and variance
\code{v}.
}
\examples{
# proposal function
ooat_rq_tau(0.2, 0.05)


# density function
ooat_dq_tau(0.3, 0.2, 0.05)
}
