% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{optimal_predict}
\alias{optimal_predict}
\title{Optimal prediction from GPsimulated object using known covariance function}
\usage{
optimal_predict(object)
}
\arguments{
\item{object}{A \code{GPsimulated} object}
}
\value{
A vector of predictions
}
\description{
Returns predictions assuming the covariance model used to simulate the data
was known
}
