% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{othermodel_predict}
\alias{othermodel_predict}
\title{Optimally predict points using an arbitrary covariance model}
\usage{
othermodel_predict(object, model)
}
\arguments{
\item{object}{A \code{GPsimulated} object}

\item{model}{A covariance model made using the \code{RandomFields} package}
}
\value{
A vector of predictions
}
\description{
To predict points from another model, such as the best-fitting Matern model,
create it using \code{RandomFields} and specify it here.
}
