% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plogspline}
\alias{plogspline}
\title{Evaluate the CDF of the log spectral density}
\usage{
plogspline(x, coefs, knots)
}
\arguments{
\item{x}{The values at which to evaluate the CDF.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}

\item{knots}{A vector of length \code{k} representing the knot locations.}
}
\value{
A vector of CDF values. These are always returned in ascending order.
This is another reason you probably don't want to use this function directly.
}
\description{
This function should probably not be used directly. It relies on there
being a (relatively) large number of \code{x} values supplied between the
boundary knots. The values of \code{x} are used to create a grid over
which the trapezoid rule is applied to determine the CDF values. If very
few \code{x} values are given, this function's behavior will be unreliable
and probably inaccurate.
}
