% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{plot.GPsimulated}
\alias{plot.GPsimulated}
\title{Plot a GPsimulated object}
\usage{
\method{plot}{GPsimulated}(x, ..., bw = FALSE)
}
\arguments{
\item{x}{A \code{GPsimulated} object}

\item{...}{Not used}

\item{bw}{\code{TRUE} if the plot should be generated in black and white. Defaults to \code{FALSE}.}
}
\description{
Plots the gridded prediction values using \code{\link[graphics]{image}}, and
draws the observation values on top with points.
}
\examples{
locations <- create_locations(50, 20, mindist = 0.002)
gp <- simulate_gp(locations, 'matern', c(1.5, 0.2, 1, 0.01))
plot(gp)
}
