% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{predict.GPsimulated}
\alias{predict.GPsimulated}
\title{Predict points from a GPsimulated object}
\usage{
\method{predict}{GPsimulated}(object, ..., beta, knots)
}
\arguments{
\item{object}{A \code{GPsimulated} object}

\item{...}{Not used}

\item{beta}{The spline coefficients}

\item{knots}{The knot locations}
}
\value{
A vector of predictions
}
\description{
Predict points from a GPsimulated object
}
