% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_natspl}
\alias{predict_natspl}
\title{Predict values from spline basis}
\usage{
predict_natspl(basis, coefs)
}
\arguments{
\item{basis}{An \code{n} by \code{k} spline basis matrix, where \code{n}
is the number of points the curve will be evaluated at and \code{k} is the
number of knots. Should be the output of \code{\link{nsbasis}}.}

\item{coefs}{A vector of length \code{k} representing the basis coefficients.}
}
\value{
A vector of length \code{n} containing the fitted values.
}
\description{
Given a natural spline basis and a set of coefficients, returns the
values along the fitted curve.
}
\seealso{
\code{\link{dloglogspline}} for the same thing but shifted so that
it integrates to 1 when exponentiated.
}
