% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{prepare_model}
\alias{prepare_model}
\title{Prepare the covariance model}
\usage{
prepare_model(family, params)
}
\arguments{
\item{family}{The parametric family the covariance function will belong to}

\item{params}{A vector of parameters for the covariance function. They must be
\itemize{
  \item \code{matern}: (nu, rho, sigma, nugget)
  \item \code{dampedcos}: (lambda, theta, sigma, nugget)
}}
}
\value{
A \code{GPmodel} object containing the following:
\itemize{
  \item \code{family}: The family
  \item \code{model}: The model, from the \code{RandomFields} package
  \item \code{params}: The parameter vector
  \item \code{specdens}: The spectral density function
  \item \code{logx}: Range of log frequency values over which
    \code{specdens} seems to be well-defined
}
}
\description{
Creates the model (from the RandomFields package) and spectral density
function for the specified covariance family.
}
\details{
This gets called internally by \code{\link{simulate_gp}}.
}
\examples{
model <- prepare_model('matern', c(1.5, 0.2, 1, 0.01))
}
