% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qlogspline}
\alias{qlogspline}
\title{Evaluate the inverse CDF of the log spectral density}
\usage{
qlogspline(p, coefs, knots, grid_size = 0.01)
}
\arguments{
\item{p}{The quantiles at which to evaluate the quantile function}

\item{coefs}{A vector of length \code{k} representing the basis coefficients}

\item{knots}{A vector of length \code{k} representing the knot locations}

\item{grid_size}{The size of the grid that the integral will be evaluated on}
}
\value{
A vector of inverse CDF values. These will always be returned in
increasing order, regardless of the original order of the quantiles.
}
\description{
AKA the quantile function.
}
