% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs_aao.R
\name{r_q_tau}
\alias{r_q_tau}
\alias{d_q_tau}
\title{The proposal distribution for the covariance scalar}
\usage{
r_q_tau(tau, v)

d_q_tau(tau, center, v)
}
\arguments{
\item{tau}{The covariance scalar}

\item{v}{The variance hyperparameter}

\item{center}{The mean of the proposal distribution}
}
\value{
For \code{r_q_tau}: A new proposed covariance scalar

For \code{d_q_tau}: The log density of tau, with respect to a
  distribution with mean \code{center}
}
\description{
The proposal distribution for the covariance scalar
}
\examples{
# proposal function
r_q_tau(0.5, 0.1)


# density function
d_q_tau(0.5, 0.4, 0.1)
}
