% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{rat_matern_t}
\alias{rat_matern_t}
\title{Ratio of Matern density to t density}
\usage{
rat_matern_t(x, nu, alpha, sigma, df)
}
\arguments{
\item{x}{a value}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{df}{degrees of freedom of the t density}
}
\value{
The ratio of the Matern density to the t density, evaluated at \code{x}
}
\description{
Ratio of Matern density to t density
}
