% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{rmatern}
\alias{rmatern}
\title{rmatern}
\usage{
rmatern(n, nu, alpha, sigma, df)
}
\arguments{
\item{n}{the number of samples to draw}

\item{nu}{smoothness parameter}

\item{alpha}{inverse range parameter}

\item{sigma}{scale parameter}

\item{df}{degrees of freedom of the envelope t density. This can be tuned to get a better acceptance rate}
}
\value{
a vector of random draws from the Matern spectral density
}
\description{
Draws random samples from a Matern spectral density.
}
