% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.R
\name{simulate_gp}
\alias{simulate_gp}
\title{Simulate a Gaussian process}
\usage{
simulate_gp(locations, family, params)
}
\arguments{
\item{locations}{A \code{GPlocations} object}

\item{family}{The parametric family of the covariance function}

\item{params}{The parameters for the covariance function}
}
\value{
A \code{GPsimulated} object. This consists of the locations object,
  with additional subobjects: \itemize{ \item \code{m}: The \code{GPmodel}
  object created by \code{\link{prepare_model}} \item \code{Y}: A vector of
  observations. These correspond to ALL locations, observed and predicted.
  Subset these according to \code{locations$locs$type}. }
}
\description{
Simulate a Gaussian process
}
\examples{
locations <- create_locations(50, 20, mindist = 0.002)
gp <- simulate_gp(locations, 'matern', c(1.5, 0.2, 1, 0.01))
}
