% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{traceplot}
\alias{traceplot}
\title{Plot the trace of spline coefficient MCMC}
\usage{
traceplot(betas, n = 4)
}
\arguments{
\item{betas}{A matrix where each row represents a set of coefficients}

\item{n}{The number of trace plots to draw. beta_1 through beta_n will be
displayed.}
}
\value{
Draws a plot. Returns \code{betas} invisibly.
}
\description{
Plot the trace of spline coefficient MCMC
}
