% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{get_cov_error_bounds}
\alias{get_cov_error_bounds}
\title{Calculate the error bounds for the covariance function}
\usage{
get_cov_error_bounds(b, x, knots)
}
\arguments{
\item{b}{A matrix in which each row is a vector of spline basis coefficients
from the MCMC}

\item{x}{A vector of x values to evaluate the error bounds over}

\item{knots}{The knots used in the MCMC}
}
\value{
The 2.5% and 97.5% percentiles at each x value
}
\description{
WARNING: this uses the GPU Monte Carlo integration and will crash your
computer if you do more than 1000 - 5000 coefficient vectors at a time.
This is because R is stupid and doesn't free up memory when it needs to.
}
\details{
TODO: make this take every 100th or so row from b and estimate the
bounds that way.
}
