% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{get_error_bounds}
\alias{get_error_bounds}
\title{Calculate the error bounds for the log-log spectral density}
\usage{
get_error_bounds(b, x, knots)
}
\arguments{
\item{b}{A matrix in which each row is a vector of spline basis coefficients
from the MCMC}

\item{x}{A vector of x values to evaluate the error bounds over}

\item{knots}{The knots used in the MCMC}
}
\value{
The 2.5% and 97.5% percentiles at each x value
}
\description{
Calculate the error bounds for the log-log spectral density
}
